use aggregations from api response
