+++ src/models/SearchResultsFilters.ts
@@ -4,8 +4,8 @@ class SearchResultsFilters {
   options: ItemAggregationOption[]
   activeFilters: string[]
   field: string
-  constructor(searchResults, field) {
-    const filterOptions = searchResults.filters.itemListElement
+  constructor(aggregationsResults, field) {
+    const filterOptions = aggregationsResults.itemListElement
     this.options = filterOptions.find((f) => f.id === field.value)?.values
     this.field = field.value
   }
